[feat] sync 액티비티 상단바, 바텀바 생성 및 원형그래프 옮김
