+++ app/src/main/java/com/example/sync_front/SplashActivity.kt
@@ -19,7 +19,7 @@ class SplashActivity : AppCompatActivity() {
         val mainExecutor: Executor = ContextCompat.getMainExecutor(this)
         backExecutor.schedule({
             mainExecutor.execute{
-                startActivity(Intent(applicationContext, ChattingActivity::class.java))
+                startActivity(Intent(applicationContext, MainActivity::class.java))
                 finish()
             }
         }, 0, TimeUnit.SECONDS)
