+++ app/src/main/res/drawable/btn_default.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <solid android:color="@color/biscay_50" />
+    <corners android:radius="12dp" />
+</shape>
