+++ app/src/main/res/drawable/btn_stroke.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <stroke
+        android:width="1dp"
+        android:color="@color/gray_10" />
+    <corners android:radius="12dp" />
+</shape>
