+++ app/src/main/res/drawable/ic_arrow_left.xml
@@ -1,0 +1,13 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <group>
+    <clip-path
+        android:pathData="M24,0l-24,0l-0,24l24,0z"/>
+    <path
+        android:pathData="M9.169,12L16.702,4.467C16.9,4.269 16.997,4.032 16.992,3.758C16.987,3.483 16.885,3.247 16.686,3.048C16.488,2.849 16.251,2.75 15.977,2.75C15.702,2.75 15.466,2.849 15.267,3.048L7.611,10.719C7.431,10.9 7.297,11.103 7.209,11.327C7.122,11.551 7.078,11.776 7.078,12C7.078,12.224 7.122,12.449 7.209,12.673C7.297,12.897 7.431,13.1 7.611,13.281L15.283,20.952C15.481,21.151 15.715,21.247 15.984,21.242C16.254,21.237 16.488,21.135 16.686,20.937C16.885,20.738 16.984,20.501 16.984,20.227C16.984,19.953 16.885,19.716 16.686,19.517L9.169,12Z"
+        android:fillColor="#555555"/>
+  </group>
+</vector>
