+++ app/src/main/res/drawable/ic_bookmark_saved.xml
@@ -1,0 +1,13 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:strokeWidth="1"
+      android:pathData="M19,21L12,16L5,21V5C5,4.47 5.211,3.961 5.586,3.586C5.961,3.211 6.47,3 7,3H17C17.53,3 18.039,3.211 18.414,3.586C18.789,3.961 19,4.47 19,5V21Z"
+      android:strokeLineJoin="round"
+      android:fillColor="#555555"
+      android:strokeColor="#555555"
+      android:strokeLineCap="round"/>
+</vector>
