+++ app/src/main/res/drawable/ic_more.xml
@@ -1,0 +1,14 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <group>
+    <clip-path
+        android:pathData="M24,0l-24,0l-0,24l24,0z"/>
+    <path
+        android:pathData="M19,14C20.105,14 21,13.105 21,12C21,10.895 20.105,10 19,10C17.895,10 17,10.895 17,12C17,13.105 17.895,14 19,14ZM14,12C14,13.105 13.105,14 12,14C10.895,14 10,13.105 10,12C10,10.895 10.895,10 12,10C13.105,10 14,10.895 14,12ZM7,12C7,13.105 6.105,14 5,14C3.895,14 3,13.105 3,12C3,10.895 3.895,10 5,10C6.105,10 7,10.895 7,12Z"
+        android:fillColor="#404040"
+        android:fillType="evenOdd"/>
+  </group>
+</vector>
