+++ app/src/main/res/drawable/ic_share.xml
@@ -1,0 +1,14 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <group>
+    <clip-path
+        android:pathData="M24,0l-24,0l-0,24l24,0z"/>
+    <path
+        android:pathData="M12.707,3.293C12.317,2.902 11.683,2.902 11.293,3.293L6.293,8.293C5.902,8.683 5.902,9.317 6.293,9.707C6.683,10.098 7.317,10.098 7.707,9.707L11,6.414V15C11,15.552 11.448,16 12,16C12.552,16 13,15.552 13,15V6.414L16.293,9.707C16.683,10.098 17.317,10.098 17.707,9.707C18.098,9.317 18.098,8.683 17.707,8.293L12.707,3.293ZM4,14C4.552,14 5,14.448 5,15V18C5,18.552 5.448,19 6,19H18C18.552,19 19,18.552 19,18V15C19,14.448 19.448,14 20,14C20.552,14 21,14.448 21,15V18C21,19.657 19.657,21 18,21H6C4.343,21 3,19.657 3,18V15C3,14.448 3.448,14 4,14Z"
+        android:fillColor="#404040"
+        android:fillType="evenOdd"/>
+  </group>
+</vector>
