+++ app/src/main/res/layout/fragment_home.xml
@@ -319,17 +319,8 @@
                     android:textColor="@color/gray_50" />
 
 
-
-
             </LinearLayout>
 
-            <com.example.sync_front.CircleGraphView
-                android:id="@+id/circle"
-                android:layout_width="match_parent"
-                android:layout_height="170dp"
-                app:layout_constraintLeft_toLeftOf="parent"
-                app:layout_constraintRight_toRightOf="parent"
-                app:layout_constraintTop_toBottomOf="@id/home_sync3" />
         </androidx.constraintlayout.widget.ConstraintLayout>
     </androidx.core.widget.NestedScrollView>
 </androidx.constraintlayout.widget.ConstraintLayout>
